refactor(firmware): add common/xr872/ shared layer, extract F7 infra
Create firmware/common/xr872/ — the XR872/XR872AT counterpart to
common/esp32/ — for SoC-generic code reusable across future XR872 screens,
and move the well-bounded infra out of the F7's monolithic main.c into it:

  - log.{c,h}        activity-log ring + POST-body accessors (hlog/hlog_len/
                     hlog_snapshot/hlog_reset), built on common/all/logbuf
  - clock.{c,h}      server-epoch software wall-clock (hokku_clock_set/now)
  - http_util.{c,h}  HTTP response-header parsing (read_resp_header_uint/str)
  - pm.{c,h}         hibernation (hokku_hibernate) with clamp constants
  - hokku_config.{c,h}  moved as-is (FDCM config store)

main.c now #includes these headers and uses hlog_len()/hlog_snapshot()
instead of touching the log ring directly. The F7 Makefile compiles
common/xr872 (added to include paths + source discovery); the host tests
include the moved .c files and add common/xr872 to their include dirs.

Screen-specific code (EPD driver, LED, battery ADC, frame-state, do_refresh/
refresh_thread orchestration, main) stays in firmware/bigme_f7/. The more-
coupled wifi/ota/net_cb/rollback/platform_init stay in main.c for a later
pass (infra-first). No behaviour change.

Verified: F7 host tests 100% pass; F7 Docker firmware build produces
xr_system.img.
