+++ firmware/bigme_f7/gcc/Makefile
@@ -45,6 +45,12 @@ INCLUDE_PATHS += -I$(PRJ_ROOT_PATH)
 HOKKU_COMMON_ALL := $(abspath $(PRJ_ROOT_PATH)/../common/all)
 INCLUDE_PATHS    += -I$(HOKKU_COMMON_ALL)
 
+# SoC-shared XR872 code (firmware/common/xr872) — reusable across XR872/XR872AT
+# screens (analogous to common/esp32 for the ESP32 boards). Compiled against the
+# real XR872 SDK here; host tests compile it separately via cmake mocks.
+HOKKU_COMMON_XR872 := $(abspath $(PRJ_ROOT_PATH)/../common/xr872)
+INCLUDE_PATHS      += -I$(HOKKU_COMMON_XR872)
+
 DIRS_IGNORE := $(PRJ_ROOT_PATH)/gcc%
 DIRS_IGNORE += $(PRJ_ROOT_PATH)/image%
 # Host-side unit tests (test_logic.c etc.) — compiled by cmake against XR872
@@ -54,10 +60,11 @@ DIRS_IGNORE += $(PRJ_ROOT_PATH)/image%
 # .c under test — linking them into the firmware would be fatal).
 DIRS_IGNORE += $(PRJ_ROOT_PATH)/test%
 DIRS_IGNORE += $(HOKKU_COMMON_ALL)/test%
+DIRS_IGNORE += $(HOKKU_COMMON_XR872)/test%
 DIRS_IGNORE += $(ROOT_PATH)/project/common/board/%
 DIRS_IGNORE += $(PRJ_BOARD)
 
-DIRS_ALL := $(shell find $(PRJ_ROOT_PATH) $(HOKKU_COMMON_ALL) $(ROOT_PATH)/project/common -type d)
+DIRS_ALL := $(shell find $(PRJ_ROOT_PATH) $(HOKKU_COMMON_ALL) $(HOKKU_COMMON_XR872) $(ROOT_PATH)/project/common -type d)
 DIRS      := $(filter-out $(DIRS_IGNORE),$(DIRS_ALL))
 DIRS      += $(PRJ_BOARD)
