+++ firmware/common/xr872/hokku_config.h
